Keyword page minor fixes
Made the contents of the page belong to class "container-fluid" (as opposed to just "container"). Additionally made the page clear table contents when search button is pressed - this means that if a user adds another search keyword, and then presses search, the current contents will clear and give space for new rows to be added (as opposed to just appending new results to the current results).