Moved update_db functionality back to db_handler
Decided that it's not good to have so much database functionality in the higher level "keywords_search.py" router. Instead, to allow progress to be shown, I passed in the websocket to the update_db function, that way the user can be notified as each youtuber is updated in the db.